Move D1-only pages into destiny1 folder
