+++ src/app/destiny1/activities/Activities.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import classNames from 'classnames';
-import BungieImage, { bungieBackgroundStyle } from '../dim-ui/BungieImage';
-import { DimStore, D1Store } from '../inventory/store-types';
-import { RootState } from '../store/reducers';
-import { sortedStoresSelector } from '../inventory/reducer';
-import SimpleCharacterTile from '../inventory/SimpleCharacterTile';
-import CollapsibleTitle from '../dim-ui/CollapsibleTitle';
-import { AppIcon, starIcon } from '../shell/icons';
+import BungieImage, { bungieBackgroundStyle } from '../../dim-ui/BungieImage';
+import { DimStore, D1Store } from '../../inventory/store-types';
+import { RootState } from '../../store/reducers';
+import { sortedStoresSelector } from '../../inventory/reducer';
+import SimpleCharacterTile from '../../inventory/SimpleCharacterTile';
+import CollapsibleTitle from '../../dim-ui/CollapsibleTitle';
+import { AppIcon, starIcon } from '../../shell/icons';
 import { t } from 'app/i18next-t';
 import _ from 'lodash';
-import { D1ManifestDefinitions } from '../destiny1/d1-definitions.service';
-import { D1StoresService } from '../inventory/d1-stores.service';
-import { DestinyAccount } from '../accounts/destiny-account.service';
-import { Loading } from '../dim-ui/Loading';
+import { D1ManifestDefinitions } from '../d1-definitions.service';
+import { D1StoresService } from '../../inventory/d1-stores.service';
+import { DestinyAccount } from '../../accounts/destiny-account.service';
+import { Loading } from '../../dim-ui/Loading';
 import { connect } from 'react-redux';
 import './activities.scss';
