+++ src/app/destiny1/loadout-builder/ExcludeItemsDropTarget.tsx
@@ -7,7 +7,7 @@ import {
   DropTargetMonitor,
   DropTargetSpec
 } from 'react-dnd';
-import { DimItem } from '../inventory/item-types';
+import { DimItem } from '../../inventory/item-types';
 
 interface ExternalProps {
   className?: string;
