+++ src/app/destiny1/loadout-builder/LoadoutBuilderLockPerk.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ArmorTypes, D1ItemWithNormalStats, LockedPerkHash, PerkCombination } from './types';
-import { D1GridNode, DimItem } from '../inventory/item-types';
-import BungieImage from '../dim-ui/BungieImage';
-import { AppIcon, plusIcon } from '../shell/icons';
+import { D1GridNode, DimItem } from '../../inventory/item-types';
+import BungieImage from '../../dim-ui/BungieImage';
+import { AppIcon, plusIcon } from '../../shell/icons';
 import LoadoutBuilderItem from './LoadoutBuilderItem';
 import _ from 'lodash';
 import LoadoutBuilderLocksDialog from './LoadoutBuilderLocksDialog';
