+++ src/app/destiny1/loadout-builder/loadout-builder.scss
@@ -1,4 +1,4 @@
-@import '../variables';
+@import '../../variables';
 
 .loadout-builder {
   margin-top: 10px;
