+++ src/app/destiny1/record-books/RecordBooks.tsx
@@ -1,22 +1,22 @@
 import React from 'react';
 import { t } from 'app/i18next-t';
 import classNames from 'classnames';
-import CollapsibleTitle from '../dim-ui/CollapsibleTitle';
-import { D1ManifestDefinitions } from '../destiny1/d1-definitions.service';
+import CollapsibleTitle from '../../dim-ui/CollapsibleTitle';
+import { D1ManifestDefinitions } from '../d1-definitions.service';
 import _ from 'lodash';
-import { count } from '../utils/util';
-import { setSetting } from '../settings/actions';
-import { D1Store } from '../inventory/store-types';
-import { storesSelector } from '../inventory/reducer';
-import { RootState } from '../store/reducers';
+import { count } from '../../utils/util';
+import { setSetting } from '../../settings/actions';
+import { D1Store } from '../../inventory/store-types';
+import { storesSelector } from '../../inventory/reducer';
+import { RootState } from '../../store/reducers';
 import { connect } from 'react-redux';
-import { Loading } from '../dim-ui/Loading';
-import { D1StoresService } from '../inventory/d1-stores.service';
-import { refresh$ } from '../shell/refresh';
-import BungieImage, { bungieBackgroundStyle } from '../dim-ui/BungieImage';
-import Objective from '../progress/Objective';
-import { DestinyAccount } from '../accounts/destiny-account.service';
-import { Subscriptions } from '../utils/rx-utils';
+import { Loading } from '../../dim-ui/Loading';
+import { D1StoresService } from '../../inventory/d1-stores.service';
+import { refresh$ } from '../../shell/refresh';
+import BungieImage, { bungieBackgroundStyle } from '../../dim-ui/BungieImage';
+import Objective from '../../progress/Objective';
+import { DestinyAccount } from '../../accounts/destiny-account.service';
+import { Subscriptions } from '../../utils/rx-utils';
 import './record-books.scss';
 
 interface ProvidedProps {
