+++ src/app/destiny1/record-books/record-books.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .record-books {
   h1 {
