+++ src/app/destiny1/routes.ts
@@ -1,10 +1,10 @@
 import { destinyAccountResolver } from '../accounts/destiny-account-resolver';
 import { ReactStateDeclaration } from '@uirouter/react';
 import Destiny from '../shell/Destiny';
-import { states as recordBookStates } from '../record-books/routes';
-import { states as activitiesStates } from '../activities/routes';
-import { states as loadoutBuilderStates } from '../d1-loadout-builder/routes';
-import { states as vendorsStates } from '../d1-vendors/routes';
+import { states as recordBookStates } from './record-books/routes';
+import { states as activitiesStates } from './activities/routes';
+import { states as loadoutBuilderStates } from './loadout-builder/routes';
+import { states as vendorsStates } from './vendors/routes';
 import Inventory from '../inventory/Inventory';
 
 // Root state for Destiny 1 views
