+++ src/app/destiny1/vendors/D1Vendors.m.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .vendors {
   composes: dim-page from global;
