+++ src/app/destiny1/vendors/D1Vendors.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { DestinyAccount } from '../accounts/destiny-account.service';
+import { DestinyAccount } from '../../accounts/destiny-account.service';
 import { UIViewInjectedProps } from '@uirouter/react';
-import { Loading } from '../dim-ui/Loading';
-import { Subscriptions } from '../utils/rx-utils';
-import { refresh$ } from '../shell/refresh';
+import { Loading } from '../../dim-ui/Loading';
+import { Subscriptions } from '../../utils/rx-utils';
+import { refresh$ } from '../../shell/refresh';
 import { dimVendorService, Vendor } from './vendor.service';
-import { D1Store } from '../inventory/store-types';
+import { D1Store } from '../../inventory/store-types';
 import _ from 'lodash';
 import D1Vendor from './D1Vendor';
 import styles from './D1Vendors.m.scss';
