+++ src/app/destiny1/vendors/vendor.service.ts
@@ -1,18 +1,18 @@
 import _ from 'lodash';
 import { get, set, keys, del } from 'idb-keyval';
-import { compareAccounts, DestinyAccount } from '../accounts/destiny-account.service';
-import { getVendorForCharacter } from '../bungie-api/destiny1-api';
-import { getDefinitions, D1ManifestDefinitions } from '../destiny1/d1-definitions.service';
-import { processItems } from '../inventory/store/d1-item-factory.service';
+import { compareAccounts, DestinyAccount } from '../../accounts/destiny-account.service';
+import { getVendorForCharacter } from '../../bungie-api/destiny1-api';
+import { getDefinitions, D1ManifestDefinitions } from '../d1-definitions.service';
+import { processItems } from '../../inventory/store/d1-item-factory.service';
 import copy from 'fast-copy';
-import { D1Store } from '../inventory/store-types';
-import { D1Item } from '../inventory/item-types';
-import { updateVendorRankings } from '../item-review/destiny-tracker.service';
-import { D1StoresService } from '../inventory/d1-stores.service';
-import { loadingTracker } from '../shell/loading-tracker';
-import { D1ManifestService } from '../manifest/d1-manifest-service';
-import { handleLocalStorageFullError } from '../compatibility';
-import store from '../store/store';
+import { D1Store } from '../../inventory/store-types';
+import { D1Item } from '../../inventory/item-types';
+import { updateVendorRankings } from '../../item-review/destiny-tracker.service';
+import { D1StoresService } from '../../inventory/d1-stores.service';
+import { loadingTracker } from '../../shell/loading-tracker';
+import { D1ManifestService } from '../../manifest/d1-manifest-service';
+import { handleLocalStorageFullError } from '../../compatibility';
+import store from '../../store/store';
 import { BehaviorSubject, ConnectableObservable, Observable } from 'rxjs';
 import { distinctUntilChanged, switchMap, publishReplay, tap, filter, map } from 'rxjs/operators';
